ai-service/knowledge_base/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.6
langchain-community==0.2.6
langchain-google-genai==1.0.6
chromadb==0.5.3
pydantic==2.7.4
kafka-python==2.0.2
